Improve splash screen support on different resolutions
Initially the splash screen assumed a HD resolution of 1920x1080. However some HDMI to VGA convertors will use a different resolution which would cause the splash screen to be cut off and hard to read on CRT monitors.
The splash screen is now scaled to fill the entire screen without preserving aspect ratio.